refactor(DVI): Re-export specials and `Matrix` from "./dvi/mod.ts"
